Hook all methods in PromiseTensor
